Reduce routes
